+++ packages/expo-camera/plugin/src/withCamera.ts
@@ -3,44 +3,12 @@ import {
   type ConfigPlugin,
   createRunOncePlugin,
   IOSConfig,
-  withProjectBuildGradle,
 } from 'expo/config-plugins';
-
-import { appendGeneratedCodeContents, type CodeMergeResults } from './appendCode';
 
 const pkg = require('expo-camera/package.json');
 
 const CAMERA_USAGE = 'Allow $(PRODUCT_NAME) to access your camera';
 const MICROPHONE_USAGE = 'Allow $(PRODUCT_NAME) to access your microphone';
-
-// Because we need the package to be added AFTER the React and Google maven packages, we create a new allprojects.
-// It's ok to have multiple allprojects.repositories, so we create a new one since it's cheaper than tokenizing
-// the existing block to find the correct place to insert our camera maven.
-const gradleMaven = [
-  `def expoCameraMavenPath = new File(["node", "--print", "require.resolve('expo-camera/package.json')"].execute(null, rootDir).text.trim(), "../android/maven")`,
-  `allprojects { repositories { maven { url(expoCameraMavenPath) } } }`,
-].join('\n');
-
-const withAndroidCameraGradle: ConfigPlugin = (config) => {
-  return withProjectBuildGradle(config, (config) => {
-    if (config.modResults.language === 'groovy') {
-      config.modResults.contents = addCameraImport(config.modResults.contents).contents;
-    } else {
-      throw new Error('Cannot add camera maven gradle because the build.gradle is not groovy');
-    }
-    return config;
-  });
-};
-
-/** @internal Exposed for testing */
-export function addCameraImport(src: string): CodeMergeResults {
-  return appendGeneratedCodeContents({
-    tag: 'expo-camera-import',
-    src,
-    generatedCode: gradleMaven,
-    comment: '//',
-  });
-}
 
 const withCamera: ConfigPlugin<
   {
@@ -66,7 +34,7 @@ const withCamera: ConfigPlugin<
     ].filter(Boolean) as string[]
   );
 
-  return withAndroidCameraGradle(config);
+  return config;
 };
 
 export default createRunOncePlugin(withCamera, pkg.name, pkg.version);